# Improve group by hash performance: avoid group-key/-state clones for hash-groupby